requirements.txt
# python==3.6.8
Django==2.0.3
djangorestframework==3.9.4
django-s3-storage==0.12.4
zappa==0.48.2
mysqlclient==1.3.12
django-cors-headers==3.0.2
PyMySQL==0.9.3